+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to this project will be documented in this file.
 * Remove duplicated calls to `IexecHubService#getTaskDescription` in `TaskManagerService`.
 ### Quality
 * Improve code quality.
+* Extract STOMP client configuration to its own bean.
 * Remove unchecked casts.
 ### Dependency Upgrades
 * Replace the deprecated `openjdk` Docker base image with `eclipse-temurin` and upgrade to Java 11.0.18 patch.